PBNetwork 구현체 위치 이동
